+++ CHANGELOG.md
@@ -6,12 +6,17 @@ All Notable changes to `Csv` will be documented in this file
 
 ### Added
 
-- None
+- `TabulatDataReader::select`
+- `TabulatDataReader::matching`
+- `TabulatDataReader::firstMatching` 
+- `TabulatDataReader::firstOrFailMatching`
+- `FragmentFinder` to implement [RFC7111](https://www.rfc-editor.org/rfc/rfc7111)
 
 ### Deprecated
 
 - Using the `$header` argument on `Statement::process` is deprecated and will be removed in
 the next version. Use `TabularDataReader::getRecords` on the returned value instead.
+It's usage will trigger a `E_USER_DEPRECATED` call.
 
 ### Fixed
 
@@ -28,7 +33,7 @@ the next version. Use `TabularDataReader::getRecords` on the returned value inst
 - `EscapeFormula::unescapeRecord` does the opposite of `EscapeFormula::escapeRecord`
 - `TabularReader::each` (implemented on the `Reader` and the `ResultSet` object)
 - `TabularReader::exists` (implemented on the `Reader` and the `ResultSet` object)
-- `TabularReader::reduce`  (implemented on the `Reader` and the `ResultSet` object)****
+- `TabularReader::reduce`  (implemented on the `Reader` and the `ResultSet` object)
 - `TabularReader::filter` (implemented on the `Reader` and the `ResultSet` object)
 - `TabularReader::slice` (implemented on the `Reader` and the `ResultSet` object)
 - `TabularReader::sorted` (implemented on the `Reader` and the `ResultSet` object)